+++ kernel/net/ipv4/tcp.c
@@ -1017,7 +1017,7 @@ ssize_t sock_tcp4_recvfrom(struct socket* sock, void* buf, size_t len, int flags
     return readable;
 }
 
-int sock_tcp4_sendto(struct socket* sock, const void *msg, size_t len, int flags, const struct sockaddr *to, socklen_t tolen)
+ssize_t sock_tcp4_sendto(struct socket* sock, const void *msg, size_t len, int flags, const struct sockaddr *to, socklen_t tolen)
 {
 #ifdef TCP_LOG_SOCK_SEND
     printk("tcp: send()\n");
