+++ kernel/net/ipv4/udp.c
@@ -240,7 +240,7 @@ ssize_t sock_udp4_recvfrom(struct socket* sock, void* buf, size_t len, int flags
     return len;
 }
 
-int sock_udp4_sendto(struct socket* sock, const void *msg, size_t len, int flags, const struct sockaddr *to, socklen_t tolen)
+ssize_t sock_udp4_sendto(struct socket* sock, const void *msg, size_t len, int flags, const struct sockaddr *to, socklen_t tolen)
 {
     struct udp4_socket_data* sock_data = (struct udp4_socket_data*) sock->data;
     struct sockaddr_in* dest_addr_in = (struct sockaddr_in*) to;
