+++ kernel/net/ipv4/udp.h
@@ -46,7 +46,7 @@ int sock_udp4_create(struct socket* sock);
 int	sock_udp4_connect(struct socket* sock, struct sockaddr* saddr, int sockaddr_len);
 int sock_udp4_bind(struct socket* sock, const struct sockaddr *addr, socklen_t addrlen);
 ssize_t sock_udp4_recvfrom(struct socket* sock, void* buf, size_t len, int flags, struct sockaddr* src_addr, socklen_t* addrlen);
-int sock_udp4_sendto(struct socket* sock, const void *msg, size_t len, int flags, const struct sockaddr *to, socklen_t tolen);
+ssize_t sock_udp4_sendto(struct socket* sock, const void *msg, size_t len, int flags, const struct sockaddr *to, socklen_t tolen);
 int sock_udp4_setsockopt(struct socket* sock, int level, int optname, const void *optval, unsigned int optlen);
 int sock_udp4_ioctl(struct socket *sock, uint64_t request, uint64_t arg);
 short sock_udp4_poll(struct socket *sock, struct file *file, struct poll_ctl *nctl);
